oh fuck me...

+++ src/graphics_new/zbuff/reset_zbuff.c
@@ -6,7 +6,7 @@
 /*   By: mfischer <mfischer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/06/09 17:55:53 by mfischer          #+#    #+#             */
-/*   Updated: 2019/06/09 18:08:04 by mfischer         ###   ########.fr       */
+/*   Updated: 2019/06/10 20:00:59 by mfischer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@ void	reset_zbuff(double value, size_t size)
 
 	buff = get_zbuff();
 	i = -1;
-	while (++i < size)
+	while ((size_t)++i < size)
 		buff[i] = value;
 }
